// client/app/api/accounting/model/AccountModel.ts
/**
 * GuanPlus.AccountingFirm.WebApi
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as models from './models';

/**
 * 科目余额Model
 */
export interface AccountModel {
    /**
     * 科目 Id
     */
    id?: string;

    /**
     * 层级
     */
    level?: string;

    /**
     * 科目编码
     */
    code?: string;

    /**
     * 父级科目编码
     */
    parentAccount?: models.IdNameModel;

    /**
     * 是否还有子科目
     */
    hasChildren?: boolean;

    /**
     * 科目名称
     */
    name?: string;

    /**
     * 科目借贷方向
     */
    balanceDirection?: models.AccountBalanceDirectionEnumModel;

    /**
     * 科目所属类型：5大类
     */
    accountType?: models.AccountTypeEnumModel;

    /**
     * 会计科目分配类型
     */
    accountAssignType?: models.AccountAssignTypeEnumModel;

    /**
     * 期初贷方金额
     */
    periodInitialCredit?: number;

    /**
     * 期初借方金额
     */
    periodInitialDebit?: number;

    /**
     * 贷方金额
     */
    credit?: number;

    /**
     * 借方金额
     */
    debit?: number;

    /**
     * 期末贷方金额
     */
    periodEndCredit?: number;

    /**
     * 期末借方金额
     */
    periodEndDebit?: number;

    /**
     * 子科目
     */
    children?: Array<models.AccountModel>;

    /**
     * 导入excel中对应的行数
     */
    importRow?: number;

    /**
     * ImportTaskId
     */
    importTaskId?: string;

}